Aggiornato tutti i campi da int a long

+++ Source/Podio .NET/Models/Item.cs
@@ -30,7 +30,7 @@ namespace PodioAPI.Models
         public string AppItemIdFormatted { get; set; }
 
         [JsonProperty("app_item_id")]
-        public int? AppItemId { get; set; }
+        public long? AppItemId { get; set; }
 
 
         [JsonProperty("created_by")]
@@ -72,7 +72,7 @@ namespace PodioAPI.Models
         public JArray References { get; set; }
 
         [JsonProperty("linked_account_id")]
-        public int? LinkedAccountId { get; set; }
+        public long? LinkedAccountId { get; set; }
 
         [JsonProperty("subscribed")]
         public bool? Subscribed { get; set; }
@@ -84,7 +84,7 @@ namespace PodioAPI.Models
         public bool IsLiked { get; set; }
 
         [JsonProperty("like_count")]
-        public int LikeCount { get; set; }
+        public long LikeCount { get; set; }
 
         [JsonProperty("app")]
         public Application App { get; set; }
@@ -114,7 +114,7 @@ namespace PodioAPI.Models
         public Grant Grant { get; set; }
 
         [JsonProperty("file_ids")]
-        public List<int> FileIds { get; set; }
+        public List<long> FileIds { get; set; }
 
         [JsonProperty("tasks")]
         public List<Task> Tasks { get; set; }
@@ -124,10 +124,10 @@ namespace PodioAPI.Models
 
         //When getting item collection
         [JsonProperty("comment_count")]
-        public int CommentCount { get; set; }
+        public long CommentCount { get; set; }
 
         [JsonProperty("task_count")]
-        public int TaskCount { get; set; }
+        public long TaskCount { get; set; }
 
         public Item()
         {
@@ -141,14 +141,14 @@ namespace PodioAPI.Models
             return fieldInstance<T>(genericField, externalId);
         }
 
-        public T Field<T>(int fieldId)
+        public T Field<T>(long fieldId)
             where T : ItemField, new()
         {
             var genericField = this.Fields.Find(field => field.FieldId == fieldId);
             return fieldInstance<T>(genericField, null, fieldId);
         }
 
-        protected T fieldInstance<T>(ItemField genericField, string externalId = null, int? fieldId = null)
+        protected T fieldInstance<T>(ItemField genericField, string externalId = null, long? fieldId = null)
             where T : ItemField, new()
         {
             T specificField = new T();
